requirements.txt
Django==6.0.3
djangorestframework
psycopg2-binary
pandas
numpy
groq
scikit-learn
gunicorn
requests
dj-database-url
httpx
pytest-django
pytest
coverage
flake8
factory-boy
faker
